+++ packages/sdk-ts/src/client/index.ts
@@ -1,5 +1,4 @@
 export * from './chain/index.js'
 export * from './indexer/index.js'
-export * from './gql/index.js'
 export * from './wasm/index.js'
 export * from './olp/index.js'