+++ packages/sdk-ts/src/core/modules/authz/msgs/MsgExec.spec.ts
@@ -1,7 +1,7 @@
-import MsgExec from './MsgExec.js'
+import MsgExec from './MsgExec'
 import { mockFactory } from '@injectivelabs/test-utils'
 import { BigNumberInBase } from '@injectivelabs/utils'
-import { MsgSend } from '../../bank/index.js'
+import MsgSend from '../../bank/msgs/MsgSend.js'
 
 const { injectiveAddress, injectiveAddress2 } = mockFactory