+++ packages/sdk-ts/src/cosmjs.ts
@@ -1,7 +1,6 @@
 import MsgTransferCosmjs from './core/modules/ibc/msgs/MsgTransferCosmjs.js'
 import { accountParser as injectiveAccountParser } from './core/accounts/AccountParser.js'
 
-/** @deprecated use @injectivelabs/sdk-ts/cosmjs */
 export * from './core/accounts/signers/index.js'
 export * from './core/stargate/index.js'
 export { MsgTransferCosmjs, injectiveAccountParser }