+++ packages/wallet-ts/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.32](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.31...@injectivelabs/wallet-ts@1.14.32) (2024-11-25)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.14.31](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.30...@injectivelabs/wallet-ts@1.14.31) (2024-11-25)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.14.30](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.22...@injectivelabs/wallet-ts@1.14.30) (2024-11-25)
+
+
+### Bug Fixes
+
+* version ([24f63c9](https://github.com/InjectiveLabs/injective-ts/commit/24f63c9d175507982ac77ce24e5cbce4cab96804))
+
+
+
+
+
+## [1.14.22](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.21...@injectivelabs/wallet-ts@1.14.22) (2024-11-25)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.14.21](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.20...@injectivelabs/wallet-ts@1.14.21) (2024-11-25)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.14.20](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.20-beta.4...@injectivelabs/wallet-ts@1.14.20) (2024-11-25)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
 ## [1.14.19](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.14.18...@injectivelabs/wallet-ts@1.14.19) (2024-11-11)
 
 **Note:** Version bump only for package @injectivelabs/wallet-ts